+++ CHANGELOG.md
@@ -2,12 +2,21 @@
 
 The complete changelog for the Costs to Expect REST API, follows the format defined at https://keepachangelog.com/en/1.0.0/
 
+## [v1.15.0] - 2019-06-xx
+### Changed
+- Category transformer updated to new setup.
+
+### Fixed
+- Collection parameters not passed through to category transformer. 
+- Header indent incorrect for v1.14.3 subheadings.
+- Description missing from /categories collection.
+
 ## [v1.14.3] - 2019-06-02
-## Added
+### Added
 - Added Twitter social card summary to the landing page.
 - Added `include-categories` and `include-subcategories` parameters to /resource-types/[resource-type]/resources/[resource]/items collection.
 
-## Changed
+### Changed
 - `postman` is now a supported value for the `source` parameter.
 
 ## [v1.14.2] - 2019-05-31